cleanup(core): refactor e2e tests for workspace
