+++ scripts/e2e-ci1.sh
@@ -7,8 +7,8 @@ mkdir -p tmp/angular
 mkdir -p tmp/nx
 
 export SELECTED_CLI=$1
-jest --maxWorkers=1 ./build/e2e/new.test.js &&
-jest --maxWorkers=1 ./build/e2e/affected.test.js &&
-jest --maxWorkers=1 ./build/e2e/affected-git.test.js &&
-jest --maxWorkers=1 ./build/e2e/command-line.test.js &&
+jest --maxWorkers=1 ./build/e2e/angular.test.js &&
+jest --maxWorkers=1 ./build/e2e/cli.test.js &&
+jest --maxWorkers=1 ./build/e2e/workspace.test.js &&
+jest --maxWorkers=1 ./build/e2e/workspace-aux-commands.test.js &&
 jest --maxWorkers=1 ./build/e2e/cypress.test.js
